ci: fix codecov intermittent failure on PRs from forked repos Forked repos do not actually have access to the codecov token, and codecov is subject to rate-limiting which leads to intermittent failures.
Making the token public is not a big deal, because we really treat all
information on codecov as informational only, and do not publicize it as
an important part of our system.

However, it’s likely that someone will report a bug bounty about it, so
I also added a quite descriptive explanation of the fact we do not care.

based on https://github.com/near/nearcore/commit/c40e43b4901715e65a4aca666fcb0485c936ecb4